requirements.txt
Flask
mysql-connector-python
pandas
matplotlib
